website: make requested changes; add index page to documentation.yml
This commit makes the requested changes to use a dedicated GitHub
action to commit and push, and to use a different permalink.
This commit also adds in the link to the index page to
documentation.yml, which was previously missed and is necessary
to have the index page be linked to by the website's sidebar.