website: Add Sphinx documentation to website using GitHub Actions (#159)
This PR adds a workflow file that builds the Sphinx documentation and
pushes it to the gem5 website. Currently, it is set to run once a week
and build documentation for the stable version of gem5. This PR also
includes a Python script that adds frontmatter and makes other
alterations to the HTML files generated by Sphinx so they are compatible
with the gem5 website.